Move ServerSynchronizationContext to its own file w/ namespace import
Per the v4 effect-smol convention (verified across Sharding,
MessageStorage, SqlClient, McpServer, etc.), services typically live in
their own module and are consumed via namespace import:

  import * as Foo from "./foo.js";
  Foo.layer        // free const Layer
  Foo.helper(...)  // free helpers
  Foo.Foo          // class for type-position uses

Move all ServerSynchronizationContext-related code (the class, layer,
guard, finalize, NoTransactionError, MultipleTransactionsError) from
src/internal/server.ts to a new src/internal/server-synchronization-context.ts.

Convert `static readonly layer` -> top-level `export const layer`. This
matches the more common v4 form (e.g. Sharding.ts:1436) and pairs with
namespace imports cleanly.

Caller updates -- all switched from named imports to namespace import:
- src/server.ts: `Effect.provide(ServerSynchronizationContext.layer)` /
  `ServerSynchronizationContext.finalize` (still spelled the same; just
  imported via `import * as ServerSynchronizationContext from ...`).
- src/server-transaction.ts: `ServerSynchronizationContext.guard`.
- src/internal/server.test.ts: `ServerSynchronizationContext.{finalize,
  guard, layer, NoTransactionError}`.

src/internal/server.ts now only holds `ServerTransactionInput` (a tag),
`OutOfOrderMutationError`, and `MutationAlreadyProcessedError`.
