+++ src/internal/server.test.ts
@@ -3,18 +3,18 @@ import { assertExitFailure } from "@effect/vitest/utils";
 import * as Cause from "effect/Cause";
 import * as Effect from "effect/Effect";
 import * as Exit from "effect/Exit";
-import { finalize, guard, NoTransactionError, ServerSynchronizationContext } from "./server.js";
+import * as ServerSynchronizationContext from "./server-synchronization-context.js";
 
 describe("ServerSynchronizationContext", () => {
   it.effect(
     "finalize should return NoTransactionError when called without guard",
     Effect.fn(function* () {
       const result = yield* Effect.void.pipe(
-        finalize,
+        ServerSynchronizationContext.finalize,
         Effect.provide(ServerSynchronizationContext.layer),
         Effect.exit,
       );
-      assertExitFailure(result, Cause.fail(new NoTransactionError()));
+      assertExitFailure(result, Cause.fail(new ServerSynchronizationContext.NoTransactionError()));
     }),
   );
 
@@ -22,8 +22,8 @@ describe("ServerSynchronizationContext", () => {
     "finalize shouldn't return errors when called with guard",
     Effect.fn(function* ({ expect }) {
       const result = yield* Effect.void.pipe(
-        guard,
-        finalize,
+        ServerSynchronizationContext.guard,
+        ServerSynchronizationContext.finalize,
         Effect.provide(ServerSynchronizationContext.layer),
         Effect.exit,
       );
